requirements.txt
selenium==3.141.0
pip==20.0.2
mysql-connector-python==8.0.19
Pillow==7.0.0
